changed HeartView changeable in the amount of the hearts

+++ BgLC/Models/IndividualsModel.swift
@@ -10,10 +10,10 @@ import Foundation
 // MARK: - IndividualsCounter
 class IndividualsModel: ObservableObject {
     @Published var name: String = ""
-    @Published var count: Double = 0.0
+    @Published var count: Int = 0
     @Published var imageName: String = ""
     
-    init(name: String, count: Double, imageName: String) {
+    init(name: String, count: Int, imageName: String) {
         self.name = name
         self.count = count
         self.imageName = imageName
